tons of changes

+++ src/main/java/gregtech/api/interfaces/tileentity/IHasWorldObjectAndCoords.java
@@ -79,13 +79,13 @@ public interface IHasWorldObjectAndCoords {
 
     Block getBlockAtSideAndDistance(ForgeDirection side, int aDistance);
 
-    byte getMetaID(int aX, int aY, int aZ);
+    int getMetaID(int aX, int aY, int aZ);
 
-    byte getMetaIDOffset(int aX, int aY, int aZ);
+    int getMetaIDOffset(int aX, int aY, int aZ);
 
-    byte getMetaIDAtSide(ForgeDirection side);
+    int getMetaIDAtSide(ForgeDirection side);
 
-    byte getMetaIDAtSideAndDistance(ForgeDirection side, int aDistance);
+    int getMetaIDAtSideAndDistance(ForgeDirection side, int aDistance);
 
     byte getLightLevel(int aX, int aY, int aZ);
